requirements.txt
streamlit
tensorflow
numpy
pillow
json5
scikit-learn
